chore: improve Badge exports
